fix: [encrypt] multiple progress dialogs are shown
the eventshandler class is singleton in single process, but while the
file chooser loaded the plugin, it's not 'singleton', the dialog and the
file manger both captured the progress/finished signal and shown the
dialogs.
it's not a good solution, the unlock logic should be split into a
seperated plugin, and the file chooser only need load it.
i left a FIXME and hope someday i will refact it.

Log: fix issue about dialogs

Bug: https://pms.uniontech.com/bug-view-237447.html